+++ examples/app-router/src/app/layout.tsx
@@ -1,5 +1,4 @@
 import "./styles.css";
-import "react18-loaders/dist/index.css";
 import { ThemeSwitcher } from "nextjs-themes";
 import { Layout } from "@repo/shared/dist/server";
 import { GlobalLoader, Header } from "@repo/shared";
@@ -7,7 +6,7 @@ import { Inter } from "next/font/google";
 import Link from "next/link";
 import { Cards, LandingPage } from "@repo/shared/dist/server";
 import { PageNavigatorCard, ThemeController } from "@repo/shared";
-import { ColorSwitch } from "nextjs-themes";
+import { ColorSwitch } from "nextjs-themes/color-switch";
 
 const inter = Inter({ subsets: ["latin"] });