+++ lib/src/hooks/use-theme.ts
@@ -7,7 +7,7 @@ import { DARK, LIGHT, SYSTEM } from "../constants";
 const colorSchemes = [SYSTEM, DARK, LIGHT] as ColorSchemeType[];
 let resolveTheme: ResolveFunc;
 
-interface UseThemeYield {
+export interface UseThemeYield {
   theme: string;
   darkTheme: string;
   lightTheme: string;
@@ -33,8 +33,6 @@ interface UseThemeYield {
  * ```tsx
  * const [] = useTheme(options);
  * ```
- *
- * @source - Source code
  */
 
 export const useTheme = (targetSelector?: string): UseThemeYield => {