+++ lib/src/server/nextjs/server-side-wrapper/server-side-wrapper.test.tsx
@@ -1,12 +1,11 @@
 import { cleanup, render, screen } from "@testing-library/react";
-import { afterEach, describe, test, beforeEach } from "vitest";
+import { afterEach, describe, test } from "vitest";
 import { NextJsSSGThemeSwitcher, ServerSideWrapper } from ".";
 
 describe("server-side-target", () => {
   afterEach(cleanup);
 
   test("test default tag", ({ expect }) => {
-    globalThis.cookies = {};
     render(<NextJsSSGThemeSwitcher />);
     expect(screen.getByTestId("server-side-target").tagName).toBe("DIV");
   });