+++ packages/shared/src/client/theme-controller/theme-selector.tsx
@@ -1,6 +1,6 @@
 "use client";
 import type { ColorSchemeType } from "nextjs-themes";
-import { useTheme } from "nextjs-themes";
+import { useTheme } from "nextjs-themes/hooks";
 import { useEffect, useMemo } from "react";
 import type { ChangeEvent } from "react";
 import { Select } from "../select";